feat: 초기 데이터 삽입
